Create HtmlStudentBio.html
Created an html file for University of Minnesota Fullstack Flex Coding Bootcamp (part time). This HTML file creates a basic template for a student bio and includes, a student's name, profile picture, a brief self description, and contact information.


Activity Description:
🏗️ Create a Webpage Using HTML

User Story:
As a student, I want to create a biography page that includes an image, a paragraph about me, and my contact information.


Acceptance Criteria:
It's done when the main header reads "Student Bio".


It's done when there is a section that includes a name, an image, and a paragraph.


It's done when there is a second section that includes a subheader labeled "Contact Info" and a list that includes links to my email, GitHub, and portfolio.